+++ docs/components/billboard.py
@@ -37,7 +37,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                 key="billboard",
                 position=[0, 0.5, 0],
             ),
-            show_helper=False,
             bgChildren=[
                 OrbitControls(key="OrbitControls")
             ],