+++ docs/components/frustum.py
@@ -43,7 +43,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                 )
                 for i in range(N)
             ],
-            show_helper=False,
             up=[0, 0, 1],
             bgChildren=[
                 OrbitControls(key="OrbitControls")