+++ docs/components/gripper.py
@@ -32,7 +32,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                 key="gripper",
                 position=[0, 0.5, 0],
             ),
-            show_helper=False,
             bgChildren=[
                 OrbitControls(key="OrbitControls")
             ],