+++ docs/components/luma_splats.py
@@ -50,7 +50,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                     position=[0, 1, 0],
                     key="luma-splats",
                 ),
-                show_helper=False,
                 up=[0, 1, 0],
                 bgChildren=[
                     OrbitControls(key="OrbitControls")