+++ docs/components/pivot.py
@@ -52,7 +52,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                 key="pivot-sphere",
             ),
 
-            show_helper=False,
             bgChildren=[OrbitControls(key="OrbitControls")],
         )