+++ docs/examples/background/depth_image.py
@@ -55,7 +55,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
             # hide the helper to only render the objects.
             up=[0, 1, 0],
             grid=False,
-            show_helper=False,
 
         )