+++ docs/tutorials/camera/frustum_transformation.py
@@ -183,7 +183,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                 ),
             ],
             grid=False,
-            show_helper=False,
             bgChildren=[
                 OrbitControls(key="OrbitControls")
             ],