+++ docs/tutorials/mujoco_interactive_simulator.py
@@ -88,7 +88,6 @@ with doc, doc.skip if MAKE_DOCS else nullcontext():
                     showCameraFrustum=True,  # Whether to show the camera  in the scene
                 ),
                 up=[0, 1, 0],
-                show_helper=False,
             ),
         )
         await sleep(2)